Reformat the equation
I remove the `\!` negative space. With the big parentheses, the space
should be as large as `\del` makes it.

For functions I prefer parentheses (), while for grouping terms I prefer
brackets []. That way it is clear whether parentheses/brackets mean
function evaluation or a plain multiplication. Inside of the
parentheses, I therefore change it into `\sbr`.

The equation is easier to read on multiple lines, so I just split that
up.

The γ-particle should be an upright `\gammaup` as it is a particle and
not a variable.

And lastly, the period after the equation should have a little spacing,
so I add a thinkspace.